==> AM-DNA-098/6_AM-DNA-098_scRNA.ipynb <==
import pandas as pd
import numpy as np
import scanpy as sc
import matplotlib as mpl
import matplotlib.pyplot as plt
from matplotlib.pyplot import rc_context
import pickle
import seaborn as sns
import scipy
import statistics as stats
from copy import copy
from ete3 import Tree
from copy import copy

sc.settings.verbosity = 0
sc.settings.figdir = '/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/scRNA/figures/'

clusterColorsFile = "/Genomics/chanlab/mchan/Adriano/TLS/TLS_TLSCL/20211102_clusterColorsTLSCL.p"
with open(clusterColorsFile,'rb') as fp:
    colorDict = pickle.load(fp)
    
greens = copy(mpl.cm.Greens)
greens.set_under("lightgray")
reds = copy(mpl.cm.Reds)
reds.set_under("lightgray")

# Load AM-DNA-097 scanpy object
TLS2 = sc.read_h5ad('/Genomics/chanlab/blaw/TLS/raw_data/scRNA/TLS12/TLS_120h_2.h5ad')
# Load the TLS timecourse scanpy object
TLS_integrated = sc.read_h5ad('/Genomics/chanlab/blaw/TLS/raw_data/scRNA/TLS12/TLS.integrated.h5ad')

# Load the TLS1 tree
TLS2_loc = '/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/2_lineage_reconstruction/AM-DNA-098_hybrid_newick_noMutationlessEdges_Labeled.nwk'
t = Tree(TLS2_loc, format=1)

# Load the allele table and lineage table
allele_table = pd.read_csv("/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/2_lineage_reconstruction/allele_table_filtered.txt", sep = '\t')
lineage_table = pd.read_csv("/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/2_lineage_reconstruction/AM-DNA-098_lineage_table.txt", sep = '\t')

# UMAPs of Marker Genes

with rc_context({'figure.figsize': (5, 5)}):
    sc.pl.umap(TLS2, color=['cell_state', 'nCount_RNA', 'nFeature_RNA'], palette = colorDict, save = '_TLS2_scRNA_QC.pdf', show = False)

genes = ['Sox3', 'Meox2', 'Cdx2', 'Tbx6', 'Ripply2', 'Kdr', 'Foxa2', 'Dppa3', 'Cer1', 'Pax6', 'Dbx1', 'Pax3', 'Pax1', 'Sox9',
         'Tcf15', 'Meox1']

for gene in genes:
    with rc_context({'figure.figsize': (5, 5)}):
        sc.pl.umap(TLS2, color=gene, title='TLS_120h_2_' + gene, use_raw = False, save = '_TLS2_' + gene + '.pdf', show = False, cmap=reds)

# Composition of TLS1 vs 120h of the Timecourse

TLS_120h = TLS_integrated[TLS_integrated.obs['orig.ident'].isin(['mGast_120h'])]

TLS2_cell_states_percent = {}
TLS_120h_cell_states_percent = {}

for cell_state in colorDict.keys():
    TLS2_cell_states_percent[cell_state] = 100 * len(TLS2[TLS2.obs['cell_state'] == cell_state]) / len(TLS2) 
    TLS_120h_cell_states_percent[cell_state] = 100 * len(TLS_120h[TLS_120h.obs['cell_state'] == cell_state]) / len(TLS_120h)

cell_percents = pd.DataFrame.from_dict([TLS_120h_cell_states_percent, TLS2_cell_states_percent])
cell_percents = cell_percents.rename(index = {0: "TLS_120h", 1: "TLS2"})

# Plot the composition vs the timecourse
colors = []
for color in cell_percents.columns:
    colors.append(colorDict[color])
        
cell_percents.plot(kind='bar', use_index = True, stacked=True, color = colors, grid = False, figsize = (5, 5))
plt.legend(loc='center left', bbox_to_anchor=(1.0, 0.5), prop={'size': 5})
plt.title('cell_state_proportions')
plt.ylabel('% cells/cluster')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/scRNA/figures/TLS2_vs_Timecourse_Cell_Percents.pdf', dpi = 300)
#plt.show()
plt.close()

# Overlay Desired Nodes onto TLS1 UMAP

# create a dictionary for the leaves in each selected node
node_dict = {}
selected_nodes = ['node24', 'node32', 'node63', 'node65']
for node in t.traverse():
    if node.name in selected_nodes:
        node_dict[node.name] = [leaf.name for leaf in node.get_leaves()]

for node in node_dict.keys():
    test = node_dict[node]
    cell_states = {}

    for state in colorDict.keys():
        cell_states[state] = 0

    for cell in test:
        cell_state = TLS2.obs.loc['TLS_120h_rep2_' + cell]['cell_state']
        cell_states[cell_state] += 1
        
    sorted_cell_states = {k: v for k, v in sorted(cell_states.items(), key=lambda item: item[1])}
    
    labels = sorted_cell_states.keys()
    sizes = sorted_cell_states.values()

    colors = []
    for label in labels:
        colors.append(colorDict[label])

    fig1, ax1 = plt.subplots()
    ax1.pie(sizes, labels=labels, autopct='%1.1f%%', colors = colors, startangle=90)
    ax1.axis('equal')  # Equal aspect ratio ensures that pie is drawn as a circle.
    plt.title(node + ' - ' + str(len(test)) + ' Cells')
    plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/scRNA/node_pies/' + node + '_composition_pie.pdf', dpi = 300)
    #plt.show()
    plt.close()

for ID in node_dict.keys():
    TLS2_node_ID = TLS2[['TLS_120h_rep2_' + s for s in node_dict[ID]]].copy()
    
    fig, ax = plt.subplots(1, 1, figsize=(5, 5))
    plt.plot(TLS2.obsm['X_umap'][:,0], TLS2.obsm['X_umap'][:,1], '.', color= "lightgray", alpha = 0.8, ms = 5, rasterized = True, markeredgewidth = 0.0)
    for cell_state in colorDict.keys():
        TLS2_node_ID_cell_state = TLS2_node_ID[TLS2_node_ID.obs['cell_state'] == cell_state].copy()
        plt.plot(TLS2_node_ID_cell_state.obsm['X_umap'][:,0], TLS2_node_ID_cell_state.obsm['X_umap'][:,1], '.', alpha = 1, ms = 10, color = colorDict[cell_state], markeredgewidth = 0.0)
    plt.axis([-15, 11, -15, 11])
    ax.set_yticks([])
    ax.set_xticks([])
    plt.title(str(ID) + " - " + str(len(node_dict[ID])) + ' Cells')
    plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/scRNA/node_umaps/' + str(ID) + '_umap_Overlay.pdf', dpi = 900)
    #plt.show()
    plt.close()

# QC of Target and Cut Sites in the scRNA seq

lineage_table["cellBC1"] = "TLS_120h_rep2_" + lineage_table["cellBC"]
lineage_table.set_index("cellBC1", inplace = True)
lineage_table.drop("cellBC", axis = 1, inplace = True)
n = lineage_table.shape[0]

aColNames = lineage_table.columns[range(0, lineage_table.shape[1], 3)].tolist()

mask = lineage_table[aColNames].isnull() == False
maskSum = mask.sum(axis = 0)

maskSum1 = maskSum / lineage_table.shape[0]
ms = maskSum1.sort_values(ascending = False)
aIntBCs = [i.split("_")[0] for i in ms.index.tolist()]

h = plt.figure(figsize = (4, 4))
plt.bar(np.arange(10), ms)
plt.ylim([0, 1.05])
plt.ylabel("fraction tS")
plt.xlabel("target site")
plt.title("AM-DNA-098_TLS2")
plt.xticks([i for i in range(len(aIntBCs))], aIntBCs, rotation = 90)
plt.subplots_adjust(left = 0.2, bottom = 0.5)
plt.savefig("/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/AM-DNA-098_tSCapture.pdf", dpi = 300)
#plt.show()
plt.close()

aIntBCs_cutSites = [i + j for i in aIntBCs for j in ["_r1","_r2","_r3"] ]

# indel histograms
maskNA = lineage_table[aIntBCs_cutSites].isnull() == True
maskNASum = maskNA.sum(axis = 0)
maskNone = lineage_table[aIntBCs_cutSites] == "[None]"
maskNoneSum = maskNone.sum(axis = 0)
maskMiss = lineage_table[aIntBCs_cutSites] == "Missing"
maskMissSum = maskMiss.sum(axis = 0)
maskTotal1 = maskNASum + maskNoneSum + maskMissSum
maskCutSum = lineage_table.shape[0] - maskTotal1

l = []
x = 0
for i in np.arange(1, len(maskMissSum) + 1):
    l.append(x)
    x = x + 1
    if i % 3 == 0:
        x = x + 1

h = plt.figure(figsize = (8,5))
ax = plt.subplot(1, 1, 1)
plt.bar(l, maskCutSum, 0.75, label ="Cut", color = "Red")
plt.bar(l, maskNoneSum, 0.75, bottom = maskCutSum + maskMissSum, label = "NoCut", color = "Grey")
plt.bar(l, maskMissSum, 0.75, bottom = maskCutSum, label = "Missing", color = "Blue")
x1 = plt.xticks(l, aIntBCs_cutSites, rotation = 90, fontsize = 8)
plt.ylabel("# cells")
plt.legend(loc = "upper right")

plt.subplots_adjust(left = 0.1, bottom = 0.3)
plt.savefig("/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/AM-DNA-098_Cut_Site_Capture.pdf", dpi = 300)
#plt.show()
plt.close()

maskPresent = lineage_table[aIntBCs_cutSites].isnull() == False
maskPresentSum = maskPresent.sum(axis = 1)
maskNone = lineage_table[aIntBCs_cutSites] == "[None]"
maskNoneSum = maskNone.sum(axis = 1)
maskMiss = lineage_table[aIntBCs_cutSites] == "Missing"
maskMissSum = maskMiss.sum(axis = 1)
maskCutSum = maskPresentSum - maskNoneSum - maskMissSum

pctCut = maskCutSum / maskPresentSum
#pctCut[0:10]

TLS2.obs = pd.concat([TLS2.obs, pctCut], axis = 1)
TLS2.obs.rename(columns={0: "pctCut"}, inplace = True)

mask = lineage_table[aColNames].isnull() == False
maskSum = mask.sum(axis = 1)
TLS2.obs["numTargetSites"] = maskSum.loc[maskSum.index]

mask = (lineage_table.isnull() == False) & (lineage_table !="[None]") & (lineage_table != "Missing")
maskSum = mask.sum(axis = 1)
TLS2.obs["numIndels"] = maskSum.loc[maskSum.index]

with rc_context({'figure.figsize': (5, 5)}):
    sc.pl.umap(TLS2, color=['GFP-Cas9-p'], cmap = greens, use_raw = False, vmin = 0, vmax = 2, legend_loc = None, 
               save = 'GFP_Expression_per_Cell.pdf', show = False)

with rc_context({'figure.figsize': (5, 5)}):
    sc.pl.umap(TLS2, color='pctCut', cmap = reds, use_raw = False, vmin = 0, vmax = 1, save = 'Percent_Cut_per_Cell.pdf', show = False)

with rc_context({'figure.figsize': (5, 5)}):
    sc.pl.umap(TLS2, color='numIndels', title = 'n Indels / Cell', cmap = reds, use_raw = False, vmin = 0, vmax = 30,
               save = 'Indels_per_Cell.pdf', show = False)

with rc_context({'figure.figsize': (5, 5)}):
    sc.pl.umap(TLS2, color='numTargetSites', cmap = reds, use_raw = False, vmin = 0, vmax = 10,
               save = 'n_Target_Sites_per_Cell.pdf', show = False)

# Histograms of Tree Metrics

GFP_expr = TLS2.X[:,21473]
GFP_expr = GFP_expr.todense()

# Add GFP positive and pctCut positive columns
TLS2.obs['GFP-Cas9-p'] = GFP_expr
TLS2.obs["GFP_pos"] = 0
x1 = TLS2.obs[TLS2.obs["GFP-Cas9-p"] > 0].index.tolist()
TLS2.obs.loc[x1, "GFP_pos"] = 1

TLS2.obs["pctCut_pos"] = 0
x1 = TLS2.obs[TLS2.obs["pctCut"] > 0].index.tolist()
TLS2.obs.loc[x1,"pctCut_pos"] = 1

len(TLS2.obs[TLS2.obs['GFP_pos'] == 1]) / len(TLS2.obs)

# plot the distribution of number of target sites for GFP negative cells
fig = plt.subplots
ax1 = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 0], x = 'numTargetSites', binwidth = 1, bins = 10)
ax1.set_xlim(0, 11)
ax1.set_xticks([0, 2, 4, 6, 8, 10])
ax1.grid(False)
ax1.set_title('GFP- {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 0])))

plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/n_target_sites_GFP-_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# plot the distribution of the number of target sites for GFP positive cells
fig = plt.subplots
ax1 = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 1], x = 'numTargetSites', binwidth = 1, bins = 10)
ax1.set_xlim(0, 11)
ax1.set_xticks([0, 2, 4, 6, 8, 10])
ax1.grid(False)
ax1.set_title('GFP+ {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 1])))

plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/n_target_sites_GFP+_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# Plot teh distribution of number of indels for GFP negative cells
fig = plt.figure()
ax = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 0], x = 'numIndels', binwidth = 1, kde = True, fill = True)
ax.set_xlim(0, 30)
ax.grid(False)
ax.set_title('GFP- {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 0])))
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/n_Indels_GFP-_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# plot the distribution of number of indels for GFP positive cells
fig = plt.figure()
ax = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 1], x = 'numIndels', binwidth = 1, kde = True, fill = True)
ax.set_xlim(0, 30)
ax.grid(False)
ax.set_title('GFP+ {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 1])))
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/n_Indels_GFP+_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# plot the distribution of pct cut in GFP negative cells
fig = plt.figure()
ax = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 0], x = 'pctCut', binwidth = 0.05, kde = True)
ax.set_xlim(0, 1)
ax.grid(False)
ax.set_title('GFP- {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 0])))
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/PctCut_GFP-_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# plot the distribution of pctcut in the GFP positive cells
fig = plt.figure()
ax = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 1], x = 'pctCut', binwidth = 0.05, kde = True)
ax.set_xlim(0, 1)
ax.grid(False)
ax.set_title('GFP+ {} Cells'.format(len(TLS2.obs[TLS2.obs['GFP_pos'] == 1])))
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/PctCut_GFP+_hist.pdf', dpi = 300)
#plt.show()
plt.close()

fig = plt.figure()
ax = sns.histplot(GFP_expr, kde = True, binwidth = 0.1, legend = False)
ax.set_xlabel('GFP-Cas9-P')
ax.set_xlim(0, 2)
plt.grid(False)
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/GFP-Cas9-P_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# Plot the histogram of GFP expression in GFP + cells
fig = plt.figure()
ax = sns.histplot(data = TLS2.obs[TLS2.obs['GFP_pos'] == 1], x = 'GFP-Cas9-p', binwidth = 0.1, kde = True)
ax.set_xlim(0, 2)
ax.grid(False)
ax.set_title('TLS2 GFP+_' + str(len(TLS2.obs[TLS2.obs['GFP_pos'] == 1]['GFP-Cas9-p'])) + '_cells')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/GFP-Cas9-P_GFP+_hist.pdf', dpi = 300)
#plt.show()
plt.close()

# Composition of GFP+ and GFP- cells

TLS2_GFPpos_cell_state_freq = {}
TLS2_GFPneg_cell_state_freq = {}

GFPpos = TLS2.obs[TLS2.obs['GFP_pos'] == 1]
GFPneg = TLS2.obs[TLS2.obs['GFP_pos'] == 0]

for cell_state in colorDict.keys():
    TLS2_GFPpos_cell_state_freq[cell_state] = 100 * len(GFPpos[GFPpos['cell_state'] == cell_state]) / len(GFPpos['cell_state'])
    TLS2_GFPneg_cell_state_freq[cell_state] = 100 * len(GFPneg[GFPneg['cell_state'] == cell_state]) / len(GFPneg['cell_state'])

cell_percents = pd.DataFrame.from_dict([TLS2_GFPpos_cell_state_freq, TLS2_GFPneg_cell_state_freq])
cell_percents = cell_percents.rename(index = {0: "GFP+", 1: "GFP-"})

colors = []
for color in cell_percents.columns:
    colors.append(colorDict[color])

cell_percents.plot(kind='bar', use_index = True, stacked=True, color = colors, grid = False)
plt.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))
plt.title('TLS2 GFP +/- Composition')
plt.ylabel('% cells/cluster')
plt.tight_layout()
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/TLS2_GFP+_vs_GFP-_Composition.pdf', dpi = 300)
#plt.show()
plt.close()

# Cell proportions of cells for GFP+/- and pctCut +/-

# % cut by cell cluster ID
cellCounts = TLS2.obs["cell_state"].value_counts()

x1 = TLS2.obs.groupby(["cell_state", "GFP_pos"]).agg({"pctCut_pos": "count"}).reset_index()
x1.rename(columns = {"pctCut_pos": "n_count"}, inplace = True)

b1 = x1[x1["GFP_pos"] == 1].reset_index(drop = True)
cellStateList = b1["cell_state"].tolist() #cellStateList
b1.set_index("cell_state",inplace = True)
b1["prop"] = b1["n_count"] / cellCounts[cellStateList]

b2 = x1[x1["GFP_pos"] == 0].reset_index(drop = True)
b2.set_index("cell_state", inplace = True)
b2["prop"] = b2["n_count"] / cellCounts[cellStateList]

x2 = TLS2.obs.groupby(["cell_state", "pctCut_pos"]).agg({"GFP_pos": "count"}).reset_index()
x2.rename(columns={"GFP_pos": "n_count"}, inplace = True)

p1 = x2[x2["pctCut_pos"] == 1].reset_index(drop = True)
p1.set_index("cell_state", inplace = True)
p1["prop"] = p1["n_count"] / cellCounts[cellStateList]

p2 = x2[x2["pctCut_pos"] == 0].reset_index(drop = True)
p2.set_index("cell_state", inplace = True)
p2["prop"] = p2["n_count"] / cellCounts[cellStateList]

# Plot the pct cut and GPF +/- for each cell state
h = plt.figure(figsize=(12,8))

ax = plt.subplot(2,2,1)
plt.bar(np.arange(len(cellStateList)),b1["n_count"],0.75,label="GFP>0",color="Green")
plt.bar(np.arange(len(cellStateList)),b2["n_count"],0.75,bottom=b1["n_count"],label="GFP=0",color="Grey")
x1 = plt.xticks(np.arange(len(cellStateList)), cellStateList, rotation=90,fontsize=8)
plt.ylabel("# cells")
plt.grid(False)
plt.legend(loc="upper right")

ax = plt.subplot(2,2,2)
plt.bar(np.arange(len(cellStateList)),b1["prop"],0.75,label="GFP>0",color="Green")
plt.bar(np.arange(len(cellStateList)),b2["prop"],0.75,bottom=b1["prop"],label="GFP=0",color="Grey")
x1 = plt.xticks(np.arange(len(cellStateList)), cellStateList, rotation=90,fontsize=8)
plt.ylabel("% cells")
plt.grid(False)
plt.legend(loc="upper right")

ax = plt.subplot(2,2,3)
plt.bar(np.arange(len(cellStateList)),p1["n_count"],0.75,label="pctCut>0",color="Red")
plt.bar(np.arange(len(cellStateList)),p2["n_count"],0.75,bottom=p1["n_count"],label="pctCut=0",color="Grey")
x1 = plt.xticks(np.arange(len(cellStateList)), cellStateList, rotation=90,fontsize=8)
plt.ylabel("# cells")
plt.grid(False)
plt.legend(loc="upper right")

ax = plt.subplot(2,2,4)
plt.bar(np.arange(len(cellStateList)),p1["prop"],0.75,label="pctCut>0",color="Red")
plt.bar(np.arange(len(cellStateList)),p2["prop"],0.75,bottom=p1["prop"],label="pctCut=0",color="Grey")
x1 = plt.xticks(np.arange(len(cellStateList)), cellStateList, rotation=90,fontsize=8)
plt.ylabel("% cells")
plt.grid(False)
plt.legend(loc="upper right")

plt.suptitle('AM-DNA-098_TLS2')
plt.subplots_adjust(left=0.1,bottom=0.2,hspace=0.5)
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/GFP_pctCut_by_cellType.pdf', dpi = 900)
#plt.show()
plt.close()

for cell_state in colorDict.keys():
    if cell_state != 'pPSM':
        plt.scatter(b1.loc[cell_state]['prop'], p1.loc[cell_state]['prop'], color = colorDict[cell_state])
    
plt.grid(False)
plt.xlim(0, 0.6)
plt.ylim(0.2, 0.8)
plt.xlabel('Proportion GFP+')
plt.ylabel('Proportion Cut')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/Cut_vs_GFP_by_Cellstate.pdf', dpi = 900)
#plt.show()
plt.close()

# boxplot of pctCut in GFP+ and GFP-
fig = plt.figure()
ax = sns.boxplot(data = TLS2.obs, x = 'GFP_pos', y = 'pctCut')
ax.set_ylim(0, 1)
ax.set_ylabel('% Cut TS/cell')
ax.grid(False)
ax.set_title('pctCut')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/PctCut_GFP+_vs_GFP-_boxplot.pdf', dpi = 900)
#plt.show()
plt.close()

# boxplot of pctCut in GFP+ and GFP-
fig = plt.figure()
ax = sns.boxplot(data = TLS2.obs, x = 'GFP_pos', y = 'pctCut')
ax.set_ylim(0, 1)
ax.set_ylabel('% Cut TS/cell')
ax.grid(False)
ax.set_title('pctCut')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/PctCut_GFP+_vs_GFP-_boxplot.pdf', dpi = 900)
#plt.show()
plt.close()

# boxplot of pctCut in GFP+ and GFP-
fig = plt.figure()
ax = sns.boxplot(data = TLS2.obs, x = 'pctCut_pos', y = 'GFP-Cas9-p')
ax.set_ylim(0, 2)
ax.set_ylabel('GFP Expression')
ax.grid(False)
ax.set_title('GFP Expression')
plt.savefig('/Genomics/chanlab/blaw/TLS/data/AM-DNA-098/lineage/3_lineage_stats/GFP_Cut_vs_Uncut_boxplot.pdf', dpi = 900)
#plt.show()
plt.close()
